custom_components/gardena_smart_system/binary_sensor.py
"""Platform setup for Gardena binary sensors."""

from homeassistant.config_entries import ConfigEntry
from homeassistant.core import HomeAssistant
from homeassistant.helpers.entity_platform import AddEntitiesCallback

from .entities.sensor import async_setup_binary_sensors


async def async_setup_entry(
    hass: HomeAssistant,
    entry: ConfigEntry,
    async_add_entities: AddEntitiesCallback,
) -> None:
    """Set up binary sensor platform."""
    await async_setup_binary_sensors(hass, entry, async_add_entities)
