requirements_dev.txt
pytest>=7.0
pytest-asyncio>=0.21
pytest-homeassistant-custom-component>=0.13
aiohttp
